app-component: move app to components folder
move App component to components folder
remove redux from app as it not necessary there
add default props
