+++ webapp/src/components/App/index.js
@@ -1,27 +1,26 @@
 import React, { Component } from 'react'
 import ReactDOM from 'react-dom'
 import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
 import classNames from 'classnames'
 
-import Content from './containers/Content'
-import Header from './components/Header'
-import Footer from './components/Footer'
+import Content from '../../containers/Content'
+import Header from '../../components/Header'
+import Footer from '../../components/Footer'
 
-import defaultLogo from './images/logo_pagarme.png'
-import style from './App.styles.css'
+import defaultLogo from '../../images/logo_pagarme.png'
+import style from './style.css'
 
 class App extends Component {
   constructor (props) {
     super(props)
 
     this.state = {
-      navigateTo: null,
+      navigateTo: 'next',
       closingEffect: false,
     }
   }
 
-  navigateTo (navigateTo) {
+  handleNavigation (navigateTo) {
     this.setState({ navigateTo })
   }
 
@@ -51,14 +50,14 @@ class App extends Component {
           <Header
             logoAlt="Pagar.me"
             logoSrc={defaultLogo}
-            onPrev={this.navigateTo.bind(this, 'prev')}
+            onPrev={this.handleNavigation.bind(this, 'prev')}
             onClose={this.close.bind(this)}
           />
           <Content navigateTo={this.state.navigateTo} />
           <Footer
             total={33000.15}
             buttonText={'Continuar'}
-            buttonClick={this.navigateTo.bind(this, 'next')}
+            buttonClick={this.handleNavigation.bind(this, 'next')}
             companyName={'Pagar.me'}
           />
         </div>
@@ -67,12 +66,12 @@ class App extends Component {
   }
 }
 
-const mapStateToProps = state => ({
-  target: state.injectedValues.configs.target,
-})
-
 App.propTypes = {
   target: PropTypes.string.isRequired,
 }
 
-export default connect(mapStateToProps)(App)
+App.defaultProps = {
+  target: 'app',
+}
+
+export default App
