+++ webapp/src/components/App/style.css
@@ -1,4 +1,4 @@
-@import "./styles/animations.css";
+@import "../../styles/animations.css";
 
 .checkout {
   position: absolute;
